toLong

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
 
     public static void main(String[] args) {
-        SalesManager unit_1= new SalesManager(new int[]{150,120,180});
+        SalesManager unit_1= new SalesManager(new long[]{150,120,180});
         System.out.println(unit_1.max());
     }
